refactor(provider): split contract ownership
